requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
astroid==2.15.6
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.9.1
blinker==1.6.2
Brotli==1.1.0
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
configparser==6.0.0
crayons==0.4.0
cryptography==41.0.3
cssselect==1.2.0
dateparser==1.1.8
demjson3==3.0.6
dill==0.3.7
exceptiongroup==1.1.3
facebook-page-scraper==5.0.2
facebook-scraper==0.2.59
fake-useragent==1.2.1
fastapi==0.103.1
frozenlist==1.4.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
iniconfig==2.0.0
isort==5.12.0
kaitaistruct==0.10
lazy-object-proxy==1.9.0
lxml==4.9.3
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.28.0
outcome==1.2.0
packaging==23.1
parse==1.19.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.3.0
pyasn1==0.5.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyee==8.2.2
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
pytest==7.4.2
pytest-timeout==2.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.8.8
requests==2.31.0
requests-html==0.10.0
selenium==4.1.0
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.27.0
tomli==2.0.1
tomlkit==0.12.1
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==1.26.16
urllib3-secure-extra==0.1.0
uvicorn==0.23.2
w3lib==2.1.2
webdriver-manager==3.2.2
websockets==10.4
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.16.2
zstandard==0.21.0
